# src/FlorenceApi/Endpoints/CaptionsEndpoints.cs
using FlorenceApi.Dtos;
using FlorenceApi.Handlers;
using Microsoft.AspNetCore.Mvc;

namespace FlorenceApi.Endpoints;

public static class CaptionsEndpoints
{
    public static IEndpointConventionBuilder MapCaptions(this IEndpointRouteBuilder app) =>
        app.MapPost("/captions", (
                [FromBody] CaptionRequest req,
                RecognitionHandler h,
                CancellationToken ct) => h.CaptionAsync(req, ct))
            .WithTags("Captions")
            .WithSummary("Generate a natural-language caption for an image.")
            .WithDescription(
                """
                Returns a single caption string describing the image content.

                The `detail` field selects verbosity:
                - `short` (default) — a brief one-sentence caption.
                - `detailed` — a longer descriptive caption with more visual detail.
                - `more_detailed` — a multi-sentence caption with fine-grained description.

                Maps to Florence-2 tasks `<CAPTION>`, `<DETAILED_CAPTION>`, `<MORE_DETAILED_CAPTION>`.
                """)
            .Accepts<CaptionRequest>("application/json")
            .Produces<CaptionResult>(StatusCodes.Status200OK)
            .ProducesProblem(StatusCodes.Status400BadRequest)
            .ProducesProblem(StatusCodes.Status413PayloadTooLarge)
            .ProducesProblem(StatusCodes.Status502BadGateway)
            .ProducesProblem(StatusCodes.Status504GatewayTimeout)
            .WithName("CreateCaption");
}